+++ src/common/row-counts.js
@@ -1,0 +1,22 @@
+function defineRowCounts()
+{
+	var defineRes = {};
+	
+	defineRes["locations"] = 359;
+	defineRes["cultures"] = 24;
+	defineRes["referralSources"] = 18;
+	defineRes["languages"] = 40;
+	defineRes["checks"] = 14;
+	defineRes["personalityTraits"] = 10;
+	defineRes["hobbies"] = 150;
+	defineRes["gaming"] = 10;
+	defineRes["allergies"] = 8;
+	defineRes["pets"] = 9;
+	defineRes["fears"] = 20;
+	defineRes["technology"] = 5;
+	defineRes["qualifications"] = 19;
+	defineRes["experienceAreas"] = 34;
+}
+
+
+module.exports = defineRowCounts();